Upgrade Fortune teller to use resource-config.json instead of running Tracing agent
