+++ fortune-demo/fortune-maven/src/main/resources/META-INF/native-image/resource-config.json
@@ -1,0 +1,7 @@
+{
+    "globs": [
+        {
+         "glob": "fortunes.u8"
+        }
+    ]
+  }
